// src/apis/model/coOrdinates.ts
/**
 * Generated by orval v6.14.3 🍺
 * Do not edit manually.
 * Farmer App API
 * OpenAPI spec version: v1
 */

export interface CoOrdinates {
  lat?: number | null;
  lng?: number | null;
  timestamp?: number;
  accuracy?: number | null;
  altitude?: number | null;
  altitudeAccuracy?: number | null;
  course?: number | null;
  speed?: number | null;
}